pyproject.toml
[tool.poetry]
name = "notwaze-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.114.1"}
uvicorn = "^0.30.6"
boto3 = "^1.35.19"
python-dotenv = "^1.0.1"
geojson = "^3.1.0"
geopandas = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
